reintepret logic moved to icicle-runtime
