+++ wrappers/rust/icicle-core/src/vec_ops/poly_vecops.rs
@@ -2,164 +2,8 @@ use super::{add_scalars, mul_scalars, scalar_mul, sub_scalars, sum_scalars, VecO
 use crate::{polynomial_ring::PolynomialRing, traits::FieldImpl};
 use icicle_runtime::{
     errors::eIcicleError,
-    memory::{DeviceSlice, HostOrDeviceSlice, HostSlice},
-    stream::IcicleStream,
+    memory::{reinterpret_slice, reinterpret_slice_mut, HostOrDeviceSlice},
 };
-use std::mem::size_of;
-
-// TODO Yuval: generalize this and move to icicle-runtime
-
-enum UnifiedSlice<'a, T: 'a> {
-    Host(&'a HostSlice<T>),
-    Device(&'a DeviceSlice<T>),
-}
-
-enum UnifiedSliceMut<'a, T: 'a> {
-    Host(&'a mut HostSlice<T>),
-    Device(&'a mut DeviceSlice<T>),
-}
-
-/// SAFETY: Caller must ensure layout of P as [P::Base; DEGREE]
-unsafe fn reinterpret_slice<From, To>(input: &(impl HostOrDeviceSlice<From> + ?Sized)) -> UnifiedSlice<'_, To>
-where
-    From: Sized,
-    To: Sized,
-{
-    let len = input.len();
-    let flat_len = len * size_of::<From>() / size_of::<To>();
-    let ptr = input.as_ptr() as *const To;
-
-    if input.is_on_device() {
-        UnifiedSlice::Device(DeviceSlice::from_raw_parts(ptr, flat_len))
-    } else {
-        UnifiedSlice::Host(HostSlice::from_raw_parts(ptr, flat_len))
-    }
-}
-
-unsafe fn reinterpret_slice_mut<From, To>(
-    input: &mut (impl HostOrDeviceSlice<From> + ?Sized),
-) -> UnifiedSliceMut<'_, To>
-where
-    From: Sized,
-    To: Sized,
-{
-    let len = input.len();
-    let flat_len = len * size_of::<From>() / size_of::<To>();
-    let ptr = input.as_mut_ptr() as *mut To;
-
-    if input.is_on_device() {
-        UnifiedSliceMut::Device(DeviceSlice::from_raw_parts_mut(ptr, flat_len))
-    } else {
-        UnifiedSliceMut::Host(HostSlice::from_raw_parts_mut(ptr, flat_len))
-    }
-}
-
-impl<'a, T> HostOrDeviceSlice<T> for UnifiedSlice<'a, T> {
-    fn is_on_device(&self) -> bool {
-        match self {
-            UnifiedSlice::Device(d) => d.is_on_device(),
-            UnifiedSlice::Host(h) => h.is_on_device(),
-        }
-    }
-
-    fn is_on_active_device(&self) -> bool {
-        match self {
-            UnifiedSlice::Device(d) => d.is_on_active_device(),
-            UnifiedSlice::Host(h) => h.is_on_active_device(),
-        }
-    }
-
-    unsafe fn as_ptr(&self) -> *const T {
-        match self {
-            UnifiedSlice::Device(d) => d.as_ptr(),
-            UnifiedSlice::Host(h) => h.as_ptr(),
-        }
-    }
-
-    unsafe fn as_mut_ptr(&mut self) -> *mut T {
-        panic!("Cannot get mutable pointer from immutable UnifiedSlice")
-    }
-
-    fn len(&self) -> usize {
-        match self {
-            UnifiedSlice::Device(d) => d.len(),
-            UnifiedSlice::Host(h) => h.len(),
-        }
-    }
-
-    fn is_empty(&self) -> bool {
-        match self {
-            UnifiedSlice::Device(d) => d.is_empty(),
-            UnifiedSlice::Host(h) => h.is_empty(),
-        }
-    }
-
-    fn memset(&mut self, _: u8, _: usize) -> Result<(), eIcicleError> {
-        panic!("Cannot memset immutable UnifiedSlice");
-    }
-
-    fn memset_async(&mut self, _: u8, _: usize, _: &IcicleStream) -> Result<(), eIcicleError> {
-        panic!("Cannot memset_async immutable UnifiedSlice");
-    }
-}
-
-impl<'a, T> HostOrDeviceSlice<T> for UnifiedSliceMut<'a, T> {
-    fn is_on_device(&self) -> bool {
-        match self {
-            UnifiedSliceMut::Device(d) => d.is_on_device(),
-            UnifiedSliceMut::Host(h) => h.is_on_device(),
-        }
-    }
-
-    fn is_on_active_device(&self) -> bool {
-        match self {
-            UnifiedSliceMut::Device(d) => d.is_on_active_device(),
-            UnifiedSliceMut::Host(h) => h.is_on_active_device(),
-        }
-    }
-
-    unsafe fn as_ptr(&self) -> *const T {
-        match self {
-            UnifiedSliceMut::Device(d) => d.as_ptr(),
-            UnifiedSliceMut::Host(h) => h.as_ptr(),
-        }
-    }
-
-    unsafe fn as_mut_ptr(&mut self) -> *mut T {
-        match self {
-            UnifiedSliceMut::Device(d) => d.as_mut_ptr(),
-            UnifiedSliceMut::Host(h) => h.as_mut_ptr(),
-        }
-    }
-
-    fn len(&self) -> usize {
-        match self {
-            UnifiedSliceMut::Device(d) => d.len(),
-            UnifiedSliceMut::Host(h) => h.len(),
-        }
-    }
-
-    fn is_empty(&self) -> bool {
-        match self {
-            UnifiedSliceMut::Device(d) => d.is_empty(),
-            UnifiedSliceMut::Host(h) => h.is_empty(),
-        }
-    }
-
-    fn memset(&mut self, value: u8, size: usize) -> Result<(), eIcicleError> {
-        match self {
-            UnifiedSliceMut::Device(d) => d.memset(value, size),
-            UnifiedSliceMut::Host(h) => h.memset(value, size),
-        }
-    }
-
-    fn memset_async(&mut self, value: u8, size: usize, stream: &IcicleStream) -> Result<(), eIcicleError> {
-        match self {
-            UnifiedSliceMut::Device(d) => d.memset_async(value, size, stream),
-            UnifiedSliceMut::Host(h) => h.memset_async(value, size, stream),
-        }
-    }
-}
 
 // <Rq,Zq> vector-mul
 pub fn polyvec_mul_by_scalar<P>(
